skyscope_os/tooling/tools_macos.py
from smolagents import tool
import os
import subprocess
import git

class MacOSPorter:
    def __init__(self, work_dir=os.path.expanduser('~/.skyscope_os/macos_porting')):
        self.work_dir = work_dir
        os.makedirs(work_dir, exist_ok=True)
        self.opensource_repos = {
            "apple_xnu": "https://github.com/apple/darwin-xnu.git",
            "opencore": "https://github.com/acidanthera/OpenCorePkg.git",
            "darling": "https://github.com/darlinghq/darling.git"
        }

    def clone_sources(self):
        """Clones the necessary open-source repositories for macOS porting."""
        for name, url in self.opensource_repos.items():
            repo_path = os.path.join(self.work_dir, name)
            if not os.path.exists(repo_path):
                print(f"Cloning {name} repo from {url}...")
                git.Repo.clone_from(url, repo_path, depth=1)
            else:
                print(f"{name} repo already cloned.")
        return "All source repositories cloned."

    def cross_compile(self, source_path: str, output_path: str, arch: str = 'x86_64') -> (bool, str):
        """Cross-compiles a C/C++ source file for macOS."""
        # This is a simplified example. A real-world scenario would require a full SDK and a more complex build system.
        clang_cmd = [
            "clang", "-target", f"{arch}-apple-darwin",
            "-o", output_path, source_path,
            "-framework", "CoreFoundation"
        ]
        try:
            result = subprocess.run(clang_cmd, capture_output=True, text=True, check=True)
            return (True, f"Compilation successful. Output at {output_path}")
        except FileNotFoundError:
            return (False, "Error: 'clang' not found. A cross-compilation toolchain is required.")
        except subprocess.CalledProcessError as e:
            return (False, f"Compilation failed: {e.stderr}")

    def sign_binary(self, binary_path: str, identity: str) -> (bool, str):
        """Signs a macOS binary. Requires being run on a macOS host with Xcode installed."""
        # This tool can only be effectively run on a macOS machine.
        if sys.platform != "darwin":
            return (False, "Error: Binary signing can only be performed on macOS.")

        sign_cmd = ["codesign", "--sign", identity, "--force", "--deep", binary_path]
        try:
            subprocess.run(sign_cmd, capture_output=True, text=True, check=True)
            return (True, f"Successfully signed {binary_path} with identity '{identity}'.")
        except FileNotFoundError:
            return (False, "Error: 'codesign' not found. Xcode Command Line Tools are required.")
        except subprocess.CalledProcessError as e:
            return (False, f"Codesign failed: {e.stderr}")


porter = MacOSPorter()

@tool
def macos_clone_sources() -> str:
    """Clones the necessary open-source repositories for macOS porting."""
    return porter.clone_sources()

@tool
def macos_cross_compile(source_path: str, output_path: str) -> str:
    """Cross-compiles a C/C++ source file for macOS."""
    success, message = porter.cross_compile(source_path, output_path)
    return message

@tool
def macos_sign_binary(binary_path: str, identity: str) -> str:
    """Signs a macOS binary. Requires being run on a macOS host."""
    success, message = porter.sign_binary(binary_path, identity)
    return message

@tool
def build_tahoe_installer_placeholder() -> str:
    """
    Placeholder for the highly complex task of building a modified macOS Tahoe installer.
    This would involve a multi-step process of downloading, patching, and assembling a custom image.
    """
    return "Placeholder: The autonomous creation of a modified macOS installer is a highly complex, research-level task. It would involve using tools like qemu-img, kpartx, and a deep understanding of the macOS boot process. This tool serves as a placeholder for that future capability."
