skyscope_os/requirements.txt
fastapi
uvicorn
prompt-toolkit
alive-progress
requests
psutil
pydantic
sqlite-utils
sentence-transformers
playwright
docker
gitpython
numpy
smolagents
evoagentx
swarms
langchain
langgraph
lief
capstone
uncompyle6
pyelftools
r2pipe
triton
jinja2
beautifulsoup4
moviepy
gtts
pyttsx3
whisper
ffmpeg-python
transformers
tqdm
torch
torchvision
torchaudio
faiss-cpu
google-auth
google-auth-oauthlib
google-api-python-client
pyyaml
arxiv
